add playlistItem
